[Refactor]: github api response 패키지 이동 (#13)
